requirements.txt
torch
pandas
python-dotenv
pyannote.audio
transformers
pydub
torchaudio
python-docx
fuzzywuzzy[speedup]
pythainlp
openai
